feat: Create reusable form to add and edit users
